requirements.txt
cloudscraper~=1.2.71
pillow~=10.4.0
pyperclip~=1.9.0
python-dotenv~=1.0.1
macos-notifications~=0.2.0